Split 'core' module into 'session' and 'patcher' modules; keep placeholder 'core' module for backwards-compatibility
