+++ tests/unit/test_patcher.py
@@ -1,4 +1,3 @@
-#!/usr/bin/env python
 from unittest.mock import patch
 
 import requests
